Compile time fixes for building without warnings on Linux/GNUstep as well as macOS in 64 bits
1) Fixed NSUInteger/NSInteger hassle which are 32 bits wide on 32 bit CPUs and 64 bits wide on 64 bit CPUs.
2) Renamed initWithOptions to initWithTreeOptions to avoid clashing with other selectors when the argument size is different on 64 bit compilations.
3) Fixed incorrect use of NSIndexSet in a test program

+++ test/CHCustomSetsTest.m
@@ -421,7 +421,7 @@ static NSArray *abc;
 	[set addObjectsFromArray:abc];
 	for (NSUInteger i = 0; i < [abc count]; i++) {
 		XCTAssertEqualObjects([set objectAtIndex:i], [abc objectAtIndex:i],
-					   @"Wrong object at index %d.", i);
+					   @"Wrong object at index %lu.", (unsigned long) i);
 	}
 }
 
@@ -515,7 +515,7 @@ static NSArray *abc;
 	XCTAssertThrows([set removeObjectAtIndex:[abc count]]);
 	for (NSInteger i = [abc count]-1; i >= 0; i--) {
 		XCTAssertEqualObjects([set lastObject], [abc objectAtIndex:i],
-							 @"Wrong object at index %d before remove.", i);
+							 @"Wrong object at index %lu before remove.", (unsigned long) i);
 		[set removeObjectAtIndex:i];
 	}
 }
